Fix problems with wxToolBar code generation (#131)
* Only look at immediate parent when create wxToolBar

The only time CreateToolBar() should be used for construction is if the immediate parent is a frame window. Otherwise, use standard new wxToolBar.

* Warn user if wxToolBar placement doesn't look right

Placing a wxToolBar under a sizer means construction won't use CreateToolBar() -- and if the sizer's parent is a frame, that's probably a mistake on the user's part, so warn about it whenever the toolbar is selected.