Clipboard (#173)
* Add OS clipboard support

Initial support for wxUiEditor, wxFormBuilder, and wxSmith.

* Always check OS clipboard before internal clipboard

* Use hash value to identify internal/external clipboard data

Hash is generated from node name and all property values of the node and all children, then stored as the  wxUiEditor format in the clipboard. as well as in MainFrame. On Paste, this value is compared between the external and internal clipboard hashes, and if they match the internal clipboard is used to prevent unnecessary memory allocation and XML parsing.

* Stop adding name to Paste menu

If the paste is coming from the OS clipboard, we won't know what the class name is until we actually paste it. Having the name show up sometimes, and not others looks really odd, so I removed it completely.

* Handle cases where parent is null