Update some Form interfaces (#264)
* Prevent exception when disabling a Form

A selected form may not have a window associated with it, so there's nothing to call enable/disable on. In this case, the Mockup panel must be recreated.

* Update properties and events for PanelForm

This will typically be created and then the class will be added as a child to another window. Since it's neither a dialog or top level window, it has it's own unique sets of properties and events.

* Changed wording for some of the help descriptions

* Add description of inherits node and base_name attribute

* Update wxDialog and wxFrame interface

Co-authored-by: Randalphwa <38287198+Randalphwa@users.noreply.github.com>