Add GetDisplayName() function
Using this function instead of the original UpdateDisplayName() means that the name can be added when the tree item is created, instead of using an empty string and then making a separate call to replace that empty string with the real one.

UpdateDisplayName() still gets called when there is a property name change, but now it just calls GetDisplayName() to get the text to set the tree item to.