Increase maximum label length
Our own ribbon panel was having problems with a width of 16 being so short that labels looked identical. Raising it to 24 before truncation takes care of that problem, and hopefully won't be so wide as to cause user problems.

Note that this is simply the limit before text is replaced with '...' -- it doesn't affect the actual label.