Change mockup size for wxRibbonBar
If the mockup size is too small, you get a question mark instead of the text for a ribbon page which affected us when we added the aui toolbar buttons. This increases the size in our own project, and adds a note about wxRibbonBar to the description of the mockup_size property.