Let user know if trying to add too many children
Instead of just saying we can't create the child node, this first checks to see if the problem is that the parent already has the maximum number of children, and if so, a message is first displayed to the user about not allowing another child of the same type. After that, the user will still get the unable to create message.