In Debug builds, add a comment with the full rc path
It gets added to base_src_includes so that it is available both in the Property Grid, and in the source code generator.

Reason for this is because in Debug, you may want to include a bunch of resource files looking for a Dialog that isn't working right. This let's you find the rc file that contained the dialog after importing.