Fix wxFB import of invalid alignment flags
This clears horizontal alignment flags if the parent is a horizontally aligned sizer, and clears vertical alignment flags if the parent is vertically aligned. It clears all alignment flags if the node has wxEXPAND set.  Without these fixes, the code will generate asserts if compiled on a debug build of wxWidgets.