Create and use new custom DirectoryProperty for type_path
This one will change directories to the current directory setting (or the directory of the project file if the property is empty). It also requires that the directory exists.

Note that all of this is part of the wxDirDialog class -- but the wxDirProperty class doesn't allow you to set the needed wxDirDialog styles.