Add ChangeEnableState function (initial commit)
Proof of concept change -- this adds a function that is called either after all the property items in a category are changed, or after any property is changed. This can be used to enable/disable property items based on the current value of a property. Each property in the category gets passed through this function.