Add support for importing wxCrafter scintilla settings
All settings except tooltip and keyword are imported. I can't figure out why anyone would want a tooltip, and we have so many properties already, it just seems like more clutter.

Adding the keywords would be doable, but the user still has to write code to actually hook them up to the specific lexer being used. I'm not seeing an advantage for the UI to add it.