Change storage for friendly/constant pairs
FontPropDlg needs to be able to iterate through the friendly names in their exact order, and std::unordered_map doesn't allow this. We also need to reverse the lookup -- given a constant, we need to find the match friendly name. Using a structure provides the std::vector of pairs, along with functions to get the vector for iteration, the constant from a name, and a name from a constant.