Support growablecols and growablerows in Mockup
These two properties were missing Mockup support in wxFlexGridSizer and wxGridBagSizer.

This also adds a check to code generation in wxFlexGridSizer so that no code is generated if a growable column or row is higher then the max number of columns or rows, then no code is generated. This prevents an ASSERT if the code is compiled as a Debug build.