Add conditions for bitmap bundle
Supporting GetBitmapBundle is going to be more complicated than I
initially thought because some functions now expect a bundle, where
others still expect a bitmap. That means we need more information from the call to determine what code to generate. For now, no one is setting
these parameters, so this is effective the same as the original code gen.