Remove prop_label from form versions of wxMenu
Both MenuBar and PopupMenu generators had a label property, but it was
never used in code generation. The PopupMenu needs something to display
in the Mockup panel, but the class name will work fine for this. In both cases
it makes more sense to display the class name in the Navigation panel
rather than expecting the user to come up with a name that is otherwise
never used.