Use single function for Images form code
Previously, this was spread across multiple portions of the
BaseCodeGenerator class. This puts all of that code into a single
function. Not only does that avoid multiple checks for the same
state, but it also avoids calls into functions that don't apply
to this form. It will also make it a *lot* easier to modify
code generation for the Images form.