Generate functions in Image form files
This is the first step in overhauling the Image form code
generator. This adds functions so that a function can be
called without needing to know the details of how the
image data is setup. This is particularly useful for the functions
that return a wxBitmapBundle, since the call doesn't need
to know how many images comprise the bundle. The user can
change the number of images at any time, and only the Image
source module will need to be recompiled.

See issues #710