Change how unsupported properties are shown
When the XRC is viewed in the Source sub-panel, unsupported
properties are added as an XML comment. These are color coded,
so it should be fairly easy to spot. These will NOT be written out
to disk.

The GenXRCInfo() no longer displays information about
unsupported properties, since these are now shown in the Source
panel.