Support minimum size in sizeritem
This also only adds variant if it is set to something other than
"normal", and proportion if it is non-zero.

The intent here is not to write out values that are the default -- that
just increases the size of the XRC file and the diffs if it is under
source management control.