Add function for add attributes to an object
The primary reason for this is to use a control's id instead of it's
variable name if the id is set to something other thant wxID_ANY.
Note that it won't matter if we are auto-generating id numbers.
While the numerical value won't be used in the XRC file, the id
string will be used to generate a unique number.