Fix for crash in MockupPreview
The problem was that AfterCreation functions were calling
GetMockup() to get the current node, as well as getting the
wxObject of child controls. This commit changes the parameters
to AfterCreation to make the node always available. It also
stores the preview wxObject* in Node, which is only available
if in a preview window.