Add support for dialogs in MockupPreview
The sizer mockups are now responsible for setting the dialog's
parent sizer (if any).

If buttons are set to default or contain one of the standard ids,
the the dialog affirmative or escape id will be set to that button.
This makes is possible to have a dialog that doesn't use
stddialogbutton, but does have a button that should close the
window -- this now works by adding a lambda to the affirmative
and escape button ids (if they exist) that will close the dialog if
the button is clicked.