Change ribbon separator icon
Only wxAuiToolBar draws a line when a separator is added, and since we
now have a toolspacer icon it makes sense to use that instead.

This commit only changes the ribbon panel separator -- toolbar separators
are a bit more complicated since the same tool is used for both wxToolBar
and wxAuiToolBar.