Switch ribbon toolbar tools to dropdown menus
Note that the drop down menus are different for the different type of
toolbars, and if the user selects the wrong one it will automatically use the
correct menu. I.e., if the user has a tool in a wxToolBar selected and chooses
the Tools dropdown from the wxAuiToolbar section of the ribbon panel, they
will still get the correct wxToolBar menu.