Remove need for ${CMAKE_CURRENT_LIST_DIR}/
In the main file list (src/files_list.cmake) using the
${CMAKE_CURRENT_LIST_DIR}/ macro in front of every file makes reading
the file list harder, as well as more work to add additional files. Combining
a macro with the directory and a transform command to CMakeLists
to add the required prefix solves the problem.

I also added spacing around the include filenames for those editors
and/or extensions that supporting opening a file based on the filename
underneath the cursor, but which don't understand parenthesis.