Expand building and licenses information
Since build/ is in .gitignore, it's fine for users to create a build/ directory
and use that in their CMake configuration, so the documentation now
explains that. It's also fine for the user to create their own setup.h file
and copy it to wxSnapshot/wx/include -- this will become more
important once we get POSIX builds working.

Since we're adding more sub modules and snapshots, it made sense to
create a bulleted list of *all* the licenses to keep the lawyers happy.