Remove checks for needed code generation
Between choosing a single language, or generating only the files in a folder,
there is no longer a way to know what the user is going to want to generate.
That means all UI used to generate the code can no longer have it's
enabled state change depending on whether all generated files are current.

This commit removes all the code that performed tests for whether
generation is need or not.