Add dialog to display generated code results
For small projects, the message box we used was sufficient. However, with
the addition of languages and multiple .cmake files and directories that
folders now enable, it's going to be a *lot* more useful to know exactly
what file(s) got updated. This new dialog is now displayed any time more
than one file got updated.