Create new language categories for forms
This breaks apart the Code Generation category into a C++ Settings,
wxPython Settings, wxLua Settings, and XRC settings with additional
properties added to all categories except C++.

I also removed the ability to create combined python and lua files. There
was no underlying support for this yet, so it's more a matter of not
supporting something that's not commonly done as far as I can tell.

Not that the only change here is UI -- there's no support yet for the new
properties.