Add additional functions to Code class
This is primarily using wxStaticText as the testbed for enhancing the
functionality of the Code class.

New line-breaking code should help a bit with extremely long lines, though
it will result in some churn for SVM like git since it will change the generated
lines.

We're not there yet, but the Code class is moving in the direction of
completely replacing gen_lang_common.h and probably a fair amount
of gen_common.h as well. Having a class where we know the node
and language certainly makes code generation briefer.