Eliminate function using bool as parameter
The std::string_view of this function was different than the std::string version
which resulted in different output depending on which was used. In addition,
the enum was being passed as a bool, and then passed to a second function
as an enum. Besides being confusing as to what is actually being passed, it
was totally reliant that no one would pass the third enum value to the
function variant that takes a string_view.