Add Python support, fix C++ bugs
If the underlined property is unchecked, then wxGenericHyperlinkCtrl must
be used for it to work on Windows. The Mockup panel was already doing
this, but the C++ code generation was not. Note that wxPython doesn't
support wxGenericHyperlinkCtrl.

There's also a bug in the wxWidgets header files -- wx/generic/hyperlink.h
will fail if wx/hyperlink.h wasn't included first, and that breaks header
alphabetical order. This commit adds a hack to force the two headers
out of order.

The XRC handler was specifying the wrong handler for this control.