Switch to namespace for Code enums
This allows callers to specify "using namespace code;" and avoid the Code::
prefix which somewhat impairs code readability. The reverse problem, of
course, is where the value is defined. However the value itself is
unimportant, and the idea is to use descriptive names so the you don't
need to know where the value is defined.