Add Python support for Images form
This generates the Images form as a Python file, and adds the necessary
import statement to all Forms that need it. It also adds the module
prefix if the image is declared externally to the current form.

Unlike C++, you *can* duplicate Python images in different forms. The only
way to prevent that is to add them to the Images form.