Regenerate for 3.1 minimum
The project file is now set to 3.1 for C++ code, so this generates conditionals
if needed. Since the C++ executable is built using wxSnapshot, it will always
be at least 3.2, so this is just to ensure that code generation is somewhat
reasonable. We don't have a way to actually run the 3.1 code.