Add import_all_dialogs property to wxFrame
If a wxFrame is used as the main window, then it will often be used to
launch dialogs directly from menu or tool events. Adding a property that
automatically imports every dialog's module makes it easier to implement
the events after the generated code instead of needing to inherit the class.