Optimize the Add() function more efficient
This avoids running any additional code unless the string starts with "wx"
or it contains a pipe character. Goal is to make this efficient to use so that
there is little if any advantage to using the Str() function.

The reason for this, is that I keep running into wxPython code generation
bugs caused by using Str() instead of Add().