Add wxGrid as second splitter window, remove OtherCtrls
All the functionality of OtherCtrls has been added to other dialogs.

Note that I did *NOT* add another wxGenericDirCtrl -- while the code for this
does work, on Windows with a networked drive, it raises an exception in the
Windows API, and depending on the network can take several minutes before
you can continue.