Add initial_enum_string to C++ form properties
This is the value for the first generated enum id in the header file. The default
is "wxID_HIGHEST + 1" but the user can set it to whatever they want. It can
be a simple numeric value, a constant defined in an external header file or
a custom offset like "wxID_HIGHEST + 1000".