Refactor list_wx_ids[]
Take advantage of C++ std::to_array and convert the list to an array. Since
inline prevents duplicates, I converted this into a header instead of a C++ file.
Finally I added an array of stock ids -- initially this will be used by IDEditorDlg,
but other classes might want access to this as well.