Initial commit to add an ID Editor dialog
Since the id property has a type, it's fairly straightforward to add this to
our property grid, which results in a button on the right side that opens
up our new dialog.

This commit just creates all the connections and events -- enough to get
the dialog displayed, but the dialog doesn't do anything yet.