Add inline to images if no translation_unit
If the image is defined in the header file, then that header file can be included
into mutliple source files which would cause the image to be defined muliple
times. Adding inline prevents the duplication, but requires a C++17 or later
compiler, so a warning comment is generated with the first image.