Refactor AutoImagesAction
This is now implemented similar to GroupUndoActions where we maintain
an array of InsertNodeAction cmds.

InsertNodeAction now supports firing a Created/Deleted event so that it's no
longer required for the newly created node to have already fired the event.
That saves having to call a ModifyPropertyAction cmd in order to change
one or more properties to non-default values.