Update wxUiEditor version
We need to synchronize the current project version number with the upcoming
beta, treating it as if the beta was a release version. That's the only way to
ensure that new properties written by any DailyBuild will mark the project
file as new so that the beta will know to warn before opening/saving the
project.