Turn off clang_analyzer in _DEBUG code block
clang_analyzer doesn't like assign values to variables that aren't used. That's
a valid warning in non-Debug builds, but sometimes making assignments
makes it easier to see what's happening while running under the debugger.
Shutting this off means the file doesn't get flagged as having a problem in
VSCode when code_analyer is turned on.