Initial Commit of Base Global ID Editor dialog
This creates the UI and their events that will allow the user to make global
changes to custom ids throughout the project. The dialog isn't hooked up
to anything, I just wanted it in the main repository so that I could work on the
inherited class in a long-running branch.

This is related to issue #1017