Display as if C++ if internal testing
Python doesn't support controls used instead of a label in wxStaticBox,
so if the user prefers Python generated code, then we don't display either
the checkbox or radio button. However, if we are using a testing version,
we need the Mockup to show us the normal C++ version, which this commit
does.