Limit description of internationalize
The _() macro does *not* call wxGetTranslation in Ruby, or at least it is not
a one-to-one change. In Python it presumably calls wx.GetTranslation, though
I am not certain of that. Since all we do is use that macro and leave it to the
language to take it from there, I reduced the description to just say it creates
a _() macro.