Fix regression in wxListView code generation
I switched to Function() so that Ruby code generation would change the
name to snake_case. However, the list item in this case is *not* a pointer,
so you can't use Function() or you end up with 'info->' instead of 'info.'
which breaks C++.

This commit special-cases the snake_case function for Ruby, and for
other languages adds the function as ".Clear(" and ".SetId(".