Added control styles to check for
DialogBlocks doesn't store property groups in the project file, so there's
no such thing as a style category. Instead, we have to check every bool
node to see if the name matches a control style. We already have code
that parses out other boolean nodes, so this is added in addition.

Related to #1180