Limit FinalImportCheck() on wxui projects
Unless this is a INTERNAL_TESTING build, loading a wxUiEditor project
should not call FinalImportCheck() since we assume the project file was
correctly written.

If INTERNAL_TESTING is set, then changes will be flagged as warnings using
variable names as the identifier, and sent to the Msgs window. Using
variable names means the testing Find Widget dialog can be used to find
the node with the change.