Short term fix for ShellRun_wx on UNIX
This entire function needs to be removed -- it's not used in wxUiEditor,
but it *is* needed by another repository that includes it. I added the wxExecute
for UNIX builds, but that doesn't work quite the same as Windows
ShellExecute. Nevertheless, this should be moved into the repository that
needs it, and removed from this one.