Add option to Code::Function
Ruby always needs wxWidgets functions converted to snake case even
if there is no preceding dot operator. I added a add_operator parameter
that defaults to true. When called with it set to false, it will fix the "wx"
prefix (if any) for non-C++ languages, and for Ruby, it convert the
function name to snake case.